:truck: 정확한 구조를 갖기 위해 상위 클래스 변경
모든 동작은 사실 엄밀히 말하자면 캔버스 내부에서 일어나는 것이 맞다. 따라서 메서드와 프로퍼티를 캔버스로 옮기고, 기존 캔버스는 확장성이 높아 상속의 방식으로 구현했다.
