+++ src/utils/math/constants.ts
@@ -1,0 +1,7 @@
+export const RADIAN = 180 / Math.PI;
+
+/**
+ * @descriptions
+ * 파이 값의 절반을 의미합니다. (PI Half의 줄임말)
+ */
+export const PIH = Math.PI / 2;
